fixed delete comment bug where updated state was re-rendering
