+++ co-make/src/components/issues/IssueComments.js
@@ -18,13 +18,27 @@ const IssueComments = ({ postId, comments, setComments }) => {
       .catch((err) => console.log(err));
   };
 
+  const deleteComment = (commentId) => {
+    axiosWithAuth()
+      .delete(`/comments/${commentId}`)
+      .then(() => {
+        fetchComments();
+      })
+      .catch((err) => console.log(err));
+  };
+
   return (
     <>
       {comments.length > 0 && (
         <div className="commentList">
           {comments.map((comment) => {
             return (
-              <Comment key={comment.id} comment={comment} postId={postId} />
+              <Comment
+                key={comment.id}
+                comment={comment}
+                postId={postId}
+                deleteComment={deleteComment}
+              />
             );
           })}
         </div>
